[Modify] updateUser action
